+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/AverageReturnPerBarCriterion.java
@@ -32,12 +32,12 @@ import org.ta4j.core.num.Num;
 /**
  * Calculates the average return per bar criterion.
  *
- * The {@link TotalProfitCriterion total profit} raised to the power of 1
+ * The {@link TotalReturnCriterion total profit} raised to the power of 1
  * divided by {@link NumberOfBarsCriterion number of bars}.
  */
-public class AverageProfitCriterion extends AbstractAnalysisCriterion {
+public class AverageReturnPerBarCriterion extends AbstractAnalysisCriterion {
 
-    private AnalysisCriterion totalProfit = new TotalProfitCriterion();
+    private AnalysisCriterion totalProfit = new TotalReturnCriterion();
 
     private AnalysisCriterion numberOfBars = new NumberOfBarsCriterion();