+++ ta4j-core/src/main/java/org/ta4j/core/tradereport/PerformanceReportGenerator.java
@@ -29,7 +29,7 @@ import org.ta4j.core.TradingRecord;
 import org.ta4j.core.analysis.criteria.ProfitLossCriterion;
 import org.ta4j.core.analysis.criteria.ProfitLossPercentageCriterion;
 import org.ta4j.core.analysis.criteria.TotalLossCriterion;
-import org.ta4j.core.analysis.criteria.TotalProfit2Criterion;
+import org.ta4j.core.analysis.criteria.TotalProfitCriterion;
 import org.ta4j.core.num.Num;
 
 /**
@@ -44,7 +44,7 @@ public class PerformanceReportGenerator implements ReportGenerator<PerformanceRe
     public PerformanceReport generate(Strategy strategy, TradingRecord tradingRecord, BarSeries series) {
         final Num totalProfitLoss = new ProfitLossCriterion().calculate(series, tradingRecord);
         final Num totalProfitLossPercentage = new ProfitLossPercentageCriterion().calculate(series, tradingRecord);
-        final Num totalProfit = new TotalProfit2Criterion().calculate(series, tradingRecord);
+        final Num totalProfit = new TotalProfitCriterion().calculate(series, tradingRecord);
         final Num totalLoss = new TotalLossCriterion().calculate(series, tradingRecord);
         return new PerformanceReport(totalProfitLoss, totalProfitLossPercentage, totalProfit, totalLoss);
     }